apps: Fix log-manager to not fail when there are empty dirs

+++ helmfile.d/charts/log-manager/files/retention.sh
@@ -23,17 +23,19 @@ TMPFILE="/tmp/lm.idx"
 
 # Define S3 functions
 s3_list_days() {
-  s3cmd --config "$S3_CONFIG" ls "s3://$S3_BUCKET/$S3_PREFIX/" | grep 'DIR' | awk '{print $2}' | sed "s#s3://$S3_BUCKET/$S3_PREFIX/##" | sed 's#/$##'
+  s3cmd --config "$S3_CONFIG" ls "s3://$S3_BUCKET/$S3_PREFIX/" | { grep 'DIR' || true; } | awk '{print $2}' | sed "s#s3://$S3_BUCKET/$S3_PREFIX/##" | sed 's#/$##'
 }
 
 s3_list_chunks() {
   S3_PATH="$1"
 
-  s3cmd --config "$S3_CONFIG" ls -r "s3://$S3_BUCKET/$S3_PREFIX/$S3_PATH/" | grep '\.gz\|\.zst' | awk '{print $4}'
+  s3cmd --config "$S3_CONFIG" ls -r "s3://$S3_BUCKET/$S3_PREFIX/$S3_PATH/" | { grep '\.gz\|\.zst' || true; } | awk '{print $4}'
 }
 
 s3_rm_chunks() {
   CHUNK_LIST="$1"
+
+  [ -s "$CHUNK_LIST" ] || return 0
 
   xargs -n1000 s3cmd --config "$S3_CONFIG" rm <"$CHUNK_LIST" >/dev/null
 }
